remote_rom: redesign packaging using size_guard
In coherence with the size_guard idea, the remote_rom packets are are
separated into a simple layer packet and a data packet. While basically
only managing common control informations, the layer packet may carry
any payload. The data packet, instead, manages the transmission of the
actual ROM payload and can be attached to the layer packet. This
simplifies the management of the notification packets UPDATE and SIGNAL
since they do not contain fields like 'offset' or 'payload_size'
anymore.

Further changes:
* Following data packets are identified as such by their offset instead
  of their packet type field. This simplifies the protocol.
* reduced redundancy in the packet building code